+++ libraries/AD5620/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2024-10-26
+- fix #3, setValue()
+- add **setPowerMode()** and **getPowerMode()**
+- fix #2, implement derived class **AD5640**
+- update unit test
+- update readme.md
+- minor edits.
+
+
 ## [0.1.0] - 2024-10-25
 - initial version, based upon AD5680